It compiles, but trying to compile the example program does not work.
The error is caused bythe GrPatLam_BasicNode case in patLamToVal, because the pattern doesn't have enough information for a tag. Then again, since the resulting GrVal is only used to compare with other GrVals, I could probably use a GrTag_Any.

Also moved the arity raising up in the transformation chain. This should be close to what we drew the diagrams for, but I forgot the actual position.