Fixed hsc2hs compilation under the cygwin. The problem with this is tricky: hsc2hs (and in general ghc) is independent of cygwin, thus it threats paths in windows style, and, forwards them in windows style to subsequent commands (eg. gcc), which in turn don't understand them. The apparent solution (found on mailing lists) is to pass "ghc" as the c compiler, but then ghc does not know about specific cygwin include directories (eg. /usr/include) which will cause big problems. The solution is to simulate the work of hsc2hs: generate the c file, compile and run it manually.
Fixed a problem with sizeof_sigset_t under Windows which didn't allow Internals.hs to compile. Although, the problem is a bit fussy: although  the function is defined only for unix systems it is used on all platform that supports tcflag_t (which may also be windows). The solution was to define it for all systems.

Still, not all libraries compile on windows.