Fixed the problem with c_stat on x64 machines. The problem was   ... do x <- c_stat ...          print x         -- prints -1          print (x == -1) -- prints False The solution was to change the return type of the equivalent c function to HsInt.
Fixed the problem with flushing IO buffers. At the (correct) termination of a program the IO buffers were not flushed and thus nothing was printed without an explicit flush . The solution was to wrap the main function inside a cleanup function which flush the buffers: (applies to standard handles: stdout, stderr).